+++ packages/finschia/src/index.ts
@@ -95,6 +95,7 @@ export {
   setupFoundationExtension,
   ThresholdDecisionPolicyEncodeObject,
 } from "./modules";
+export { NodeExtension, setupNodeExtension } from "./modules";
 export { IbcExtension, setupIbcExtension } from "./modules";
 export { stakingplusTypes } from "./modules";
 export {