+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to
 ## [Unreleased]
 
 ### Added
+- [\#61](https://github.com/line/lbmjs/pull/61) @lbmjs/finschia: Add query minimum gas price
 - [\#58](https://github.com/line/lbmjs/pull/58) @lbmjs/finschia: Add test foundation amino
 ### Changed